Move DefaultRouter to router module
